refactor: auth, member 관련 api path managers -> members 변경
