+++ backend/src/main/java/com/woowacourse/zzimkkong/controller/MemberController.java
@@ -4,9 +4,7 @@ import com.woowacourse.zzimkkong.config.logaspect.LogMethodExecutionTime;
 import com.woowacourse.zzimkkong.domain.LoginEmail;
 import com.woowacourse.zzimkkong.dto.member.*;
 import com.woowacourse.zzimkkong.dto.member.oauth.OauthMemberSaveRequest;
-import com.woowacourse.zzimkkong.dto.member.LoginUserEmail;
 import com.woowacourse.zzimkkong.service.MemberService;
-import com.woowacourse.zzimkkong.service.PresetService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
@@ -21,22 +19,20 @@ import static com.woowacourse.zzimkkong.dto.ValidatorMessage.EMPTY_MESSAGE;
 
 @LogMethodExecutionTime(group = "controller")
 @RestController
-@RequestMapping("/api/managers")
+@RequestMapping("/api/members")
 @Validated
 public class MemberController {
     private final MemberService memberService;
-    private final PresetService presetService;
 
-    public MemberController(final MemberService memberService, final PresetService presetService) {
+    public MemberController(final MemberService memberService) {
         this.memberService = memberService;
-        this.presetService = presetService;
     }
 
     @PostMapping
     public ResponseEntity<Void> join(@RequestBody @Valid final MemberSaveRequest memberSaveRequest) {
         MemberSaveResponse memberSaveResponse = memberService.saveMember(memberSaveRequest);
         return ResponseEntity
-                .created(URI.create("/api/managers/" + memberSaveResponse.getId()))
+                .created(URI.create("/api/members/" + memberSaveResponse.getId()))
                 .build();
     }
 
@@ -44,7 +40,7 @@ public class MemberController {
     public ResponseEntity<Void> joinByOauth(@RequestBody @Valid final OauthMemberSaveRequest oauthMemberSaveRequest) {
         MemberSaveResponse memberSaveResponse = memberService.saveMemberByOauth(oauthMemberSaveRequest);
         return ResponseEntity
-                .created(URI.create("/api/managers/" + memberSaveResponse.getId()))
+                .created(URI.create("/api/members/" + memberSaveResponse.getId()))
                 .build();
     }
 
@@ -55,31 +51,6 @@ public class MemberController {
             @Email(message = EMAIL_MESSAGE) final String email) {
         memberService.validateDuplicateEmail(email);
         return ResponseEntity.ok().build();
-    }
-
-    @PostMapping("/presets")
-    public ResponseEntity<Void> createPreset(
-            @RequestBody @Valid final PresetCreateRequest presetCreateRequest,
-            @LoginEmail final LoginUserEmail loginUserEmail) {
-        PresetCreateResponse presetCreateResponse = presetService.savePreset(presetCreateRequest, loginUserEmail);
-        return ResponseEntity
-                .created(URI.create("/api/managers/presets/" + presetCreateResponse.getId()))
-                .build();
-    }
-
-    @GetMapping("/presets")
-    public ResponseEntity<PresetFindAllResponse> findAllPresets(@LoginEmail final LoginUserEmail loginUserEmail) {
-        PresetFindAllResponse presetFindAllResponse = presetService.findAllPresets(loginUserEmail);
-        return ResponseEntity.ok().body(presetFindAllResponse);
-    }
-
-    @DeleteMapping("/presets/{presetId}")
-    public ResponseEntity<Void> deletePreset(
-            @PathVariable final Long presetId,
-            @LoginEmail final LoginUserEmail loginUserEmail) {
-        presetService.deletePreset(presetId, loginUserEmail);
-
-        return ResponseEntity.noContent().build();
     }
 
     @GetMapping("/me")
