+++ backend/src/test/java/com/woowacourse/zzimkkong/controller/AuthControllerTest.java
@@ -129,10 +129,10 @@ class AuthControllerTest extends AcceptanceTest {
         return RestAssured
                 .given(getRequestSpecification()).log().all()
                 .accept("application/json")
-                .filter(document("member/login", getRequestPreprocessor(), getResponsePreprocessor()))
+                .filter(document("members/login", getRequestPreprocessor(), getResponsePreprocessor()))
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(loginRequest)
-                .when().post("/api/managers/login/token")
+                .when().post("/api/members/login/token")
                 .then().log().all().extract();
     }
 
@@ -140,9 +140,9 @@ class AuthControllerTest extends AcceptanceTest {
         return RestAssured
                 .given(getRequestSpecification()).log().all()
                 .accept("application/json")
-                .filter(document("member/login/oauth/" + oauthProvider.name(), getRequestPreprocessor(), getResponsePreprocessor()))
+                .filter(document("members/login/oauth/" + oauthProvider.name(), getRequestPreprocessor(), getResponsePreprocessor()))
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/api/managers/" + oauthProvider + "/login/token?code=" + code)
+                .when().get("/api/members/" + oauthProvider + "/login/token?code=" + code)
                 .then().log().all().extract();
     }
 
@@ -151,9 +151,9 @@ class AuthControllerTest extends AcceptanceTest {
                 .given(getRequestSpecification()).log().all()
                 .accept("application/json")
                 .header("Authorization", token)
-                .filter(document("member/token/" + docName, getRequestPreprocessor(), getResponsePreprocessor()))
+                .filter(document("members/token/" + docName, getRequestPreprocessor(), getResponsePreprocessor()))
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/api/managers/token")
+                .when().post("/api/members/token")
                 .then().log().all().extract();
     }
 }
